bug fixing, Preventing the "Sum of the background spots and the total gene counts not equal for specific label images" assertion error. The wSpotCell array however is declared as np.float64, ie it takes more memory now

+++ pciSeq/src/core/main.py
@@ -201,7 +201,7 @@ class VarBayes:
         nN = self.nN
         nS = self.spots.data.gene_name.shape[0]
 
-        wSpotCell = np.zeros([nS, nN], dtype=np.float32)
+        wSpotCell = np.zeros([nS, nN], dtype=np.float64)
         gn = self.spots.data.gene_name.values
         expected_counts = self.single_cell.log_mean_expression.loc[gn].values
         logeta_bar = self.genes.logeta_bar[self.spots.gene_id]
